Refactor: move function to other file.
